azure-function/requirements.txt
azure-functions
azure-identity
azure-storage-blob
azure-storage-queue
openai
pdfplumber
pyodbc
requests
tiktoken
